requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# 🤖 Aether Multi-Agent Bot Dependencies
# AI & ML Libraries
anthropic>=0.8.0
google-generativeai>=0.3.0
groq>=0.4.0
openai>=1.0.0

# Telegram Bot Framework
aiogram>=3.0.0
aiohttp>=3.8.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Database & Storage
redis>=4.5.0
sqlalchemy>=2.0.0
alembic>=1.10.0

# Async HTTP Client
httpx>=0.24.0
aiofiles>=23.0.0

# Security & Cryptography
cryptography>=41.0.0
python-jose>=3.3.0
passlib>=1.7.0

# Monitoring & Logging
prometheus-client>=0.16.0
structlog>=23.0.0

# Development & Testing
pytest>=7.2.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Environment & Configuration
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
click>=8.1.0
typer>=0.9.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0

# Security Tools
bandit>=1.7.5
safety>=2.3.0
pip-audit>=2.6.0

# Session Management
itsdangerous>=2.1.0
secure-cookie>=0.4.0

# File Processing
python-multipart>=0.0.6
Pillow>=10.1.0

# Background Tasks
celery[redis]>=5.3.0
kombu>=5.3.0

# Data Validation
email-validator>=2.1.0
phonenumbers>=8.13.0

# Rate Limiting
slowapi>=0.1.9
limits>=3.7.0

# Logging
loguru>=0.7.0
python-json-logger>=2.0.0

# Configuration
pyyaml>=6.0.1
toml>=0.10.2

# Health Checks
health-check>=0.3.0
psutil>=5.9.0

# Time & Dates
python-dateutil>=2.8.0
pytz>=2023.3

# JSON Processing
orjson>=3.9.0
ujson>=5.8.0

# Caching
cachetools>=5.3.0
diskcache>=5.6.0

# Webhooks
webhooks>=0.8.0
requests>=2.31.0

# Environment
python-decouple>=3.8
environs>=10.0.0

# API Documentation
fastapi-users>=12.1.0
fastapi-pagination>=0.12.0

# Database ORM (optional)
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# WebSocket (if needed)
websockets>=12.0

# Task Queue
dramatiq[redis]>=1.15.0

# Monitoring Client
statsd>=4.0.0
datadog>=0.47.0

# Testing Tools
httpx>=0.25.0
factory-boy>=3.3.0
faker>=20.1.0

# Performance
uvloop>=0.19.0
orjson>=3.9.0

# Security Headers
secure-headers>=0.5.0
